Project structure (#46)
* Separate Client and ConnectionManager in different files

* Change RoomData and ActiveRooms data structures

* Import from server instead of current package

* Use list as condition

This uses the list as a condition instead of checking if the length of the list is 0.
